mpi tutorial

+++ universe_render/snap_io.py
@@ -4,7 +4,7 @@ def from_npy(filename):
     """
     This is the data format used in the example. 
     """
-    data = np.load(filename).T
+    data = np.load(filename).T.astype(float)
     n_part = data.shape[0]
     hsml = data[:,5]
     rho = data[:,4]
